pyproject.toml
[tool.poetry]
name = "task2-fastapi"
version = "0.1.0"
description = ""
authors = ["ponOlgap <ponOlgap@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
async-exit-stack = "^1.0.1"
async-generator = "^1.10"
python-multipart = "^0.0.6"
certifi = "^2023.11.17"
pylint = "^3.0.2"
psycopg2 = "^2.9.9"
alembic = "^1.12.1"
asyncpg = "^0.29.0"
aiosqlite = "^0.19.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
pydantic-settings = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
